ci: Add util to test structure to clean cloudv2 after testrun
  To minimize code duplication, util is using the same RpCloudApiClient
  class that is used in redpanda_cloud class. Also, to isolate it from
  ducktape runtime, its placed in tests folder to utilize rptest as an
  external module and prevent any CI-related modifications.
